+++ run_tests.py
@@ -1,4 +1,9 @@
 #!/usr/bin/env python
+
+# Copyright 2012 splinter authors. All rights reserved.
+# Use of this source code is governed by a BSD-style
+# license that can be found in the LICENSE file.
+
 # -*- coding: utf-8 -*-
 import argparse
 import sys
@@ -16,6 +21,7 @@ from tests.fake_webapp import start_flask_app, EXAMPLE_APP
 
 parser = argparse.ArgumentParser('Run splinter tests')
 parser.add_argument('-w', '--which', action='store')
+parser.add_argument('-f', '--failfast', action='store_true')
 
 
 class Env(object):
@@ -79,13 +85,19 @@ def get_modules(modules_str):
     return modules
 
 
-def run_suite(suite):
+def get_result(args):
     result = unittest.TextTestResult(sys.stdout, descriptions=True, verbosity=1)
+
+    if args.failfast:
+        result.failfast = True
+
+    return result
+
+
+def run_suite(suite, result):
     suite.run(result)
 
     sys.stdout.write("\n\n")
-
-    return result
 
 
 def get_suite_from_modules(modules):
@@ -134,7 +146,8 @@ if __name__ == '__main__':
     else:
         suite = get_complete_suite()
 
-    result = run_suite(suite)
+    result = get_result(args)
+    run_suite(suite, result)
     print_failures(result)
     print_errors(result)
     sys.stdout.write("%d tests. %d failures. %d errors.\n\n" % (result.testsRun, len(result.failures), len(result.errors)))