+++ splinter/driver/webdriver/chrome.py
@@ -1,5 +1,9 @@
 # -*- coding: utf-8 -*-
 
+# Copyright 2012 splinter authors. All rights reserved.
+# Use of this source code is governed by a BSD-style
+# license that can be found in the LICENSE file.
+
 from selenium.webdriver import Chrome
 from selenium.webdriver.chrome.options import Options
 from splinter.driver.webdriver import BaseWebDriver, WebDriverElement
@@ -7,6 +11,9 @@ from splinter.driver.webdriver.cookie_manager import ChromeCookieManager
 
 
 class WebDriver(BaseWebDriver):
+
+    driver_name = "Chrome"
+
     def __init__(self, user_agent=None):
         self._patch_subprocess()
         options = Options()
@@ -22,9 +29,3 @@ class WebDriver(BaseWebDriver):
         self._cookie_manager = ChromeCookieManager(self.driver)
 
         super(WebDriver, self).__init__()
-
-    def attach_file(self, name, value):
-        """
-        Chrome doesn't have support for file uploading.
-        """
-        raise NotImplementedError