+++ splinter/driver/webdriver/firefox.py
@@ -1,5 +1,9 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
+
+# Copyright 2012 splinter authors. All rights reserved.
+# Use of this source code is governed by a BSD-style
+# license that can be found in the LICENSE file.
+
 import subprocess
 
 from selenium.webdriver import Firefox
@@ -10,7 +14,9 @@ from splinter.driver.webdriver.cookie_manager import CookieManager
 
 class WebDriver(BaseWebDriver):
 
-    def __init__(self, profile=None, extensions=None, user_agent=None):
+    driver_name = "Firefox"
+
+    def __init__(self, profile=None, extensions=None, user_agent=None, profile_preferences=None):
         self.old_popen = subprocess.Popen
         firefox_profile = FirefoxProfile(profile)
         firefox_profile.set_preference('extensions.logging.enabled', False)
@@ -18,6 +24,10 @@ class WebDriver(BaseWebDriver):
 
         if user_agent is not None:
             firefox_profile.set_preference('general.useragent.override', user_agent)
+
+        if profile_preferences:
+            for key, value in profile_preferences.iteritems():
+                firefox_profile.set_preference(key, value)
 
         if extensions:
             for extension in extensions: