+++ splinter/exceptions.py
@@ -1,12 +1,17 @@
 # -*- coding: utf-8 -*-
 
+# Copyright 2012 splinter authors. All rights reserved.
+# Use of this source code is governed by a BSD-style
+# license that can be found in the LICENSE file.
+
+
 class DriverNotFoundError(Exception):
     """
     Exception raised when a driver is not found.
 
     Example:
 
-        >>> from splinter.browser import Browser
+        >>> from splinter import Browser
         >>> b = Browser('unknown driver') # raises DriverNotFoundError
     """
     pass