+++ splinter/request_handler/status_code.py
@@ -1,5 +1,10 @@
 # -*- coding: utf-8 -*-
 
+# Copyright 2012 splinter authors. All rights reserved.
+# Use of this source code is governed by a BSD-style
+# license that can be found in the LICENSE file.
+
+
 class HttpResponseError(Exception):
     """
     Represents an HTTP response error.
@@ -35,6 +40,9 @@ class StatusCode(object):
             return 0
         return 1
 
+    def __str__(self):
+        return "%s - %s" % (self.code, self.reason)
+
     def is_valid_response(self):
         """
         Returns ``True`` if the response is valid (:attr:`code` < 400).
@@ -48,7 +56,7 @@ class StatusCode(object):
 
     def is_success(self):
         """
-        Returns ``True`` if the response was succed, otherwise, returns ``False``.
+        Returns ``True`` if the response was succeed, otherwise, returns ``False``.
         """
         if self.code not in self.http_errors:
             return True