+++ docs/news.rst
@@ -1,3 +1,7 @@
+.. Copyright 2012 splinter authors. All rights reserved.
+   Use of this source code is governed by a BSD-style
+   license that can be found in the LICENSE file.
+
 .. meta::
     :description: New splinter features on earlier versions.
     :keywords: splinter, python, news, documentation, tutorial, web application
@@ -8,6 +12,9 @@ splinter news
 
 See below the changes for each splinter release.
 
+- :doc:`what's new in splinter 0.4.8 </news/0.4.8>`
+- :doc:`what's new in splinter 0.4.7 </news/0.4.7>`
+- :doc:`what's new in splinter 0.4.4.1 </news/0.4.4.1>`
 - :doc:`what's new in splinter 0.4.4 </news/0.4.4>`
 - :doc:`what's new in splinter 0.4.3 </news/0.4.3>`
 - :doc:`what's new in splinter 0.4.2 </news/0.4.2>`